requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pytesseract>=0.3.10
Pillow>=10.1.0
python-dateutil>=2.8.2
jinja2>=3.1.2
pdfplumber>=0.10.3
pandas>=2.1.0